requirements.txt
pymilvus
langchain
openai==0.28
tiktoken
fastapi
uvicorn
pymongo
pypdf
python-multipart
azure-storage-blob